+++ modules/ssm/cloudtrail.tf
@@ -1,0 +1,109 @@
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudtrail
+resource "aws_cloudtrail" "ssm_api_calls" {
+  name                          = "ssm-api-calls-trail"
+  s3_bucket_name                = var.log_bucket.id
+  s3_key_prefix                 = "cloudtrail/ssm"
+  include_global_service_events = false
+  is_multi_region_trail         = false
+  enable_logging                = true
+  kms_key_id                    = aws_kms_key.ssm.arn
+
+  event_selector {
+    read_write_type           = "All"
+    include_management_events = true
+  }
+
+  insight_selector {
+    insight_type = "ApiCallRateInsight"
+  }
+
+  cloud_watch_logs_group_arn = "${aws_cloudwatch_log_group.ssm_cloudtrail.arn}:*"
+  cloud_watch_logs_role_arn  = aws_iam_role.cloudtrail_logs.arn
+
+  tags = var.tags
+}
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudwatch_log_group
+resource "aws_cloudwatch_log_group" "ssm_cloudtrail" {
+  name              = "devops/aws/cloudtrail/ssm"
+  retention_in_days = var.log_retention_days
+  kms_key_id        = aws_kms_key.ssm.arn
+
+  tags = var.tags
+}
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/iam_role
+resource "aws_iam_role" "cloudtrail_logs" {
+  name               = "ssm-cloudtrail-logs-role"
+  assume_role_policy = data.aws_iam_policy_document.cloudtrail_assume_role.json
+
+  tags = var.tags
+}
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/data-sources/iam_policy_document
+data "aws_iam_policy_document" "cloudtrail_assume_role" {
+  statement {
+    effect = "Allow"
+    principals {
+      type        = "Service"
+      identifiers = ["cloudtrail.amazonaws.com"]
+    }
+    actions = ["sts:AssumeRole"]
+  }
+}
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/iam_role_policy
+resource "aws_iam_role_policy" "cloudtrail_logs" {
+  name   = "ssm-cloudtrail-logs-policy"
+  role   = aws_iam_role.cloudtrail_logs.id
+  policy = data.aws_iam_policy_document.cloudtrail_logs.json
+}
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/data-sources/iam_policy_document
+data "aws_iam_policy_document" "cloudtrail_logs" {
+  statement {
+    effect = "Allow"
+    actions = [
+      "logs:CreateLogStream",
+      "logs:PutLogEvents"
+    ]
+    resources = ["${aws_cloudwatch_log_group.ssm_cloudtrail.arn}:*"]
+  }
+}
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/s3_bucket_policy
+resource "aws_s3_bucket_policy" "cloudtrail_bucket_policy" {
+  bucket = var.log_bucket.id
+  policy = data.aws_iam_policy_document.cloudtrail_bucket_policy.json
+}
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/data-sources/iam_policy_document
+data "aws_iam_policy_document" "cloudtrail_bucket_policy" {
+  statement {
+    sid    = "AWSCloudTrailAclCheck"
+    effect = "Allow"
+    principals {
+      type        = "Service"
+      identifiers = ["cloudtrail.amazonaws.com"]
+    }
+    actions   = ["s3:GetBucketAcl"]
+    resources = [var.log_bucket.arn]
+  }
+
+  statement {
+    sid    = "AWSCloudTrailWrite"
+    effect = "Allow"
+    principals {
+      type        = "Service"
+      identifiers = ["cloudtrail.amazonaws.com"]
+    }
+    actions   = ["s3:PutObject"]
+    resources = ["${var.log_bucket.arn}/cloudtrail/ssm/AWSLogs/${data.aws_caller_identity.current.account_id}/*"]
+    condition {
+      test     = "StringEquals"
+      variable = "s3:x-amz-acl"
+      values   = ["bucket-owner-full-control"]
+    }
+  }
+}