+++ modules/ssm/ssm_policy.tf
@@ -1,8 +1,4 @@
 
-locals {
-  log_name = "devops/aws/ssm/sessions"
-}
-
 # To Delete:
 # aws ssm delete-document --name SSM-SessionManagerRunShell
 #
@@ -43,13 +39,4 @@ resource "aws_ssm_document" "sessions" {
     aws_kms_key.ssm,
     aws_kms_alias.ssm
   ]
-}
-
-# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloudwatch_log_group
-resource "aws_cloudwatch_log_group" "ssm_sessions" {
-  name              = local.log_name
-  kms_key_id        = aws_kms_key.ssm.arn
-  skip_destroy      = false
-  retention_in_days = var.log_retention_days
-  tags              = var.tags
 }